Feat: 포인트 사용 API 구현

+++ src/main/java/com/musinsa_payments/exam/common/constant/PointPolicyConstant.java
@@ -5,7 +5,7 @@ public class PointPolicyConstant {
      * 최대 적립 가능 포인트의 기본값 (1회 적립 시 10만 포인트까지 가능).
      * DB에 정책 값이 없을 경우 사용할 기본 최대 적립 포인트 값입니다.
      */
-    public static final int DEFAULT_MAX_ACCUMULATE_POINT = 100000;
+    public static final long DEFAULT_MAX_ACCUMULATE_POINT = 100000;
 
     /**
      * 사용자가 보유할 수 있는 최대 포인트의 기본값 (20만 포인트).
